[URECA-55] fix: 약관페이지에서 새로고침할 경우 마이페이지로 넘어가는 오류 수정
