+++ src/app/(policy)/policy/_components/policy-agree/index.tsx
