+++ src/app/(policy)/policy/_components/policy-view/index.tsx
